+++ CHANGELOG.md
@@ -7,7 +7,7 @@ PrestaShop 1.6.x and 1.7.x
 
 ## [3.3.7] - 2022-01-03
 ### Added
-- Add a button to trigger the sync of the terminals with the gateway
+- Enable possibility to synchronize terminals based on store country with a Button in Prestashop
 
 ## [3.3.6] - 2021-07-05
 ### Fixed