+++ CHANGELOG.md
@@ -5,6 +5,10 @@ PrestaShop 1.6.x and 1.7.x
 
 # Changelog
 
+## [3.3.7] - 2022-01-03
+### Added
+- Enable possibility to synchronize terminals based on store country with a Button in Prestashop
+
 ## [3.3.6] - 2021-07-05
 ### Fixed
 - Add configuration to enable/disable CVV field